Use http provider so we can better inspect errors
