+++ rust/Cargo.toml
@@ -8,7 +8,6 @@ members = [
     "agents/validator",
     "chains/hyperlane-ethereum",
     "chains/hyperlane-fuel",
-    "ethers-fallback",
     "ethers-prometheus",
     "gelato",
     "hyperlane-base",
