+++ rust/chains/hyperlane-ethereum/src/lib.rs
@@ -14,7 +14,9 @@ use hyperlane_core::*;
 pub use retrying::{RetryingProvider, RetryingProviderError};
 
 #[cfg(not(doctest))]
-pub use crate::{interchain_gas::*, mailbox::*, multisig_ism::*, provider::*, trait_builder::*};
+pub use crate::{
+    fallback::*, interchain_gas::*, mailbox::*, multisig_ism::*, provider::*, trait_builder::*,
+};
 
 #[cfg(not(doctest))]
 mod tx;
@@ -44,6 +46,9 @@ mod contracts;
 
 /// Retrying Provider
 mod retrying;
+
+/// Fallback provider
+mod fallback;
 
 /// Ethereum connection configuration
 #[derive(Debug, serde::Deserialize, Clone)]
