+++ rust/config/test/test_config.json
@@ -49,7 +49,7 @@
         "url": ""
       },
       "index": {
-        "from": "20"
+        "from": "21"
       }
     }
   },
